# PERF/CLN: to_csv