--- a/pandas/io/formats/csvs.py
+++ b/pandas/io/formats/csvs.py
@@ -160,7 +160,7 @@ class CSVFormatter:
 
     def _initialize_chunksize(self, chunksize: int | None) -> int:
         if chunksize is None:
-            return (100000 // (len(self.cols) or 1)) or 1
+            return 100000 // len(self.cols) or 1
         return int(chunksize)
 
     @property